requirements.txt
aioredis==2.0.1
alembic==1.9.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
awscli==1.34.8
bcrypt==3.2.0
billiard>=4.2.0
botocore==1.35.8
build==1.2.1
cachetools==5.5.0
celery==5.3.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.0
dill==0.3.6
dnslib==0.9.25
dnspython==2.6.1
docutils==0.16
email-validator
ecdsa==0.19.0
fastapi==0.109.1
fastapi-limiter==0.1.6
h11==0.14.0
httptools==0.6.1
idna==3.8
importlib_metadata==8.4.0
jmespath==1.0.1
kombu==5.4.0
localstack==3.7.0
localstack-core==3.7.0
localstack-ext==3.7.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.1
passlib[bcrypt]
plux==1.11.1
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.3
pyaes==1.6.1
pyasn1==0.6.0
pycparser==2.22
pydantic
pydantic_core==2.20.1
Pygments==2.18.0
pyotp==2.9.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
rich==13.8.0
rsa==4.7.2
s3transfer==0.10.2
semver==3.0.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.35.0
tabulate==0.9.0
tailer==0.4.1
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.19
uvicorn==0.22.0
uvloop==0.20.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
zipp==3.20.1